fixed some typos and add some safe control

+++ ebe_scripts/average_event_spvn.py
@@ -34,8 +34,15 @@ try:
     working_folder = path.abspath(argv[1])
     avg_folder = path.join(path.abspath(argv[2]),
                            working_folder.split('/')[-1])
+    print("output folder: %s" % avg_folder)
     if(path.isdir(avg_folder)):
-        shutil.rmtree(avg_folder)
+        print("folder %s already exists!" % avg_folder)
+        var = raw_input("do you want to delete it? [y/N]")
+        if 'y' in var:
+            shutil.rmtree(avg_folder)
+        else:
+            print("please choose another folder path~")
+            exit(0)
     mkdir(avg_folder)
 except IndexError:
     print("Usage: average_event_spvn.py working_folder results_folder")
@@ -129,7 +136,7 @@ def calculate_chi_62(vn_array):
 def calculate_chi_63(vn_array):
     v6_array = vn_array[:, 5]
     v3_array = vn_array[:, 2]
-    nev = len(v2_array)
+    nev = len(v3_array)
     chi_63_num = v6_array*((v3_array.conjugate())**2)
     chi_63_den = (abs(v3_array))**4                  
     chi_63_num_ave = mean(chi_63_num)
@@ -599,10 +606,10 @@ for ipart, particle_id in enumerate(particle_list):
         chi_62_alice, chi_62_err_alice = calculate_chi_62(vn_alice_array2)
         chi_63_alice, chi_63_err_alice = calculate_chi_63(vn_alice_array2)
         # print "cms non-linear response coefficents"
-        chi_4_cms, chi_4_cms = calculate_chi_4(vn_cms_array2)
-        chi_5_cms, chi_5_cms = calculate_chi_5(vn_cms_array2)
-        chi_62_cms, chi_62_cms = calculate_chi_62(vn_cms_array2)
-        chi_63_cms, chi_63_cms = calculate_chi_63(vn_cms_array2)
+        chi_4_cms, chi_4_err_cms = calculate_chi_4(vn_cms_array2)
+        chi_5_cms, chi_5_err_cms = calculate_chi_5(vn_cms_array2)
+        chi_62_cms, chi_62_err_cms = calculate_chi_62(vn_cms_array2)
+        chi_63_cms, chi_63_err_cms = calculate_chi_63(vn_cms_array2)
 
     # calculate vn distribution for charged hadrons
     if particle_id == '9999':
@@ -687,7 +694,7 @@ for ipart, particle_id in enumerate(particle_list):
         # output non-linear response coefficients chi_n for cms pt cut
         output_filename = ("non_linear_response_coefficients_CMS.dat")
         f = open(output_filename, 'w')
-        f.write("type  value  stat. err\n")
+        f.write("# type  value  stat. err\n")
         f.write("4  %.10e  %.10e\n" % (chi_4_cms, chi_4_err_cms))
         f.write("5  %.10e  %.10e\n" % (chi_5_cms, chi_5_err_cms))
         f.write("62  %.10e  %.10e\n" % (chi_62_cms, chi_62_err_cms))
@@ -696,9 +703,9 @@ for ipart, particle_id in enumerate(particle_list):
         shutil.move(output_filename, avg_folder)
 
         # output non-linear response coefficients chi_n for alice pt cut
-        output_filename = ("%s_non_linear_response_coefficients_ALICE.dat")
+        output_filename = ("non_linear_response_coefficients_ALICE.dat")
         f = open(output_filename, 'w')
-        f.write("type  value  stat. err\n")
+        f.write("# type  value  stat. err\n")
         f.write("4  %.10e  %.10e\n" % (chi_4_alice, chi_4_err_alice))
         f.write("5  %.10e  %.10e\n" % (chi_5_alice, chi_5_err_alice))
         f.write("62  %.10e  %.10e\n" % (chi_62_alice, chi_62_err_alice))
